int로 되있던 자료형 모두 수정

+++ TrickyTowerDefense/Assets/Scripts/Tower/Attackable.cs
@@ -8,5 +8,5 @@ public abstract class Attackable : MonoBehaviour
     protected float attackCoolTime = 0f;
     protected float attackTimer = 0f;
 
-    public abstract void Attack(int damage, EnemyScript target);
+    public abstract void Attack(float damage, EnemyScript target);
 }
